KEYCLOAK-11691 Broker tests re-structure
