+++ testsuite/integration-arquillian/tests/base/src/test/java/org/keycloak/testsuite/account/AccountBrokerTest.java
@@ -51,9 +51,6 @@ import static org.keycloak.testsuite.admin.ApiUtil.resetUserPassword;
 public class AccountBrokerTest extends AbstractBaseBrokerTest {
 
     @Page
-    protected LoginPage loginPage;
-
-    @Page
     protected AccountFederatedIdentityPage identityPage;
 
     @Override
