+++ testsuite/integration-arquillian/tests/base/src/test/java/org/keycloak/testsuite/broker/KcOidcBrokerParameterForwardTest.java
@@ -16,7 +16,7 @@ import org.keycloak.representations.idm.UserRepresentation;
 import org.keycloak.testsuite.Assert;
 import org.keycloak.testsuite.arquillian.SuiteContext;
 
-public class KcOidcBrokerParameterForwardTest extends KcOidcBrokerTest {
+public class KcOidcBrokerParameterForwardTest extends AbstractBrokerTest {
 
     private static final String FORWARDED_PARAMETER = "forwarded_parameter";
     private static final String FORWARDED_PARAMETER_VALUE = "forwarded_value";
@@ -48,7 +48,7 @@ public class KcOidcBrokerParameterForwardTest extends KcOidcBrokerTest {
         driver.navigate().to(driver.getCurrentUrl() + queryString);
 
         log.debug("Clicking social " + bc.getIDPAlias());
-        accountLoginPage.clickSocial(bc.getIDPAlias());
+        loginPage.clickSocial(bc.getIDPAlias());
 
         waitForPage(driver, "log in to", true);
 
@@ -64,7 +64,7 @@ public class KcOidcBrokerParameterForwardTest extends KcOidcBrokerTest {
         Assert.assertThat("\"" + PARAMETER_NOT_FORWARDED +"\"" + " should be NOT part of the url",
                 driver.getCurrentUrl(), not(containsString(PARAMETER_NOT_FORWARDED)));
 
-        accountLoginPage.login(bc.getUserLogin(), bc.getUserPassword());
+        loginPage.login(bc.getUserLogin(), bc.getUserPassword());
         waitForPage(driver, "update account information", false);
 
         updateAccountInformationPage.assertCurrent();
